[rename] assets 로고 폴더 이동
