+++ src/app/(main)/(components)/Footer.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import logoVertical4 from '@/assets/logo_vertical_4.png';
+import logoVertical4 from '@/assets/logo/logo_vertical_4.png';
 import Link from 'next/link';
 import useMultilingual from '@/utils/useMultilingual';
 import { useAtom } from 'jotai';
