+++ src/constant/locales/home/news-section.ts
@@ -2,10 +2,6 @@ const NEWS_SECTION_STRINGS = {
   KOREA_IS_NOW: {
     ko: '한국은 지금',
     en: 'Korea is now'
-  },
-  MORE: {
-    ko: '더보기',
-    en: 'MORE'
   }
 };
